Move DenormalizationResult to Result namespace
