+++ src/Symfony/Component/Serializer/Serializer.php
@@ -29,6 +29,7 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+use Symfony\Component\Serializer\Result\DenormalizationResult;
 
 /**
  * Serializer serializes and deserializes data.
